+++ tests/test_components/test_medium.py
@@ -312,14 +312,12 @@ def test_gain_medium(log_capture):
     assert_log_level(log_capture, "WARNING")
 
 
-def test_medium2d():
+def test_medium2d(log_capture):
     sigma = 0.45
     thickness = 0.01
     cond_med = td.Medium(conductivity=sigma)
     medium = td.Medium2D.from_medium(cond_med, thickness=thickness)
 
-    _ = medium.plot(freqs=[2e14, 3e14], ax=AX)
-    plt.close()
     _ = medium.plot_sigma(freqs=[2e14, 3e14], ax=AX)
     plt.close()
     assert np.isclose(medium.ss.to_medium().conductivity, sigma * thickness, rtol=RTOL)
@@ -337,6 +335,22 @@ def test_medium2d():
         sigma / 3,
         rtol=RTOL,
     )
+
+    # 2d geometry should raise a warning, but 3d should not
+    td.Structure(medium=medium3d, geometry=td.Box(size=(1, 1, 1)))
+
+    # no warnings so far
+    assert_log_level(log_capture, None)
+
+    # these should give warnings
+    td.Structure(medium=medium3d, geometry=td.Box(size=(1, 0, 1)))
+    assert_log_level(log_capture, "WARNING")
+
+    log_capture.clear()
+
+    _ = medium.plot(freqs=[2e14, 3e14], ax=AX)
+    plt.close()
+    assert_log_level(log_capture, "WARNING")
 
 
 def test_rotation():