+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Bug in `Simulation.eps_bounds` that was always setting the lower bound to 1.
 - Extended lower limit of frequency range for `Graphene` to zero.
+- Improved warnings for `Medium2D`.
 
 ## [2.4.0rc2] - 2023-8-21